SQL GPT/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Database
sqlalchemy==2.0.23

# LLM Integration
openai==2.8.1

# Utilities
python-dotenv==1.0.0
sqlparse==0.4.4

# Optional: For better performance
aiofiles==23.2.1
